+++ CHANGELOG.md
@@ -9,7 +9,9 @@ All notable changes to the package will be documented here.
 ## [4.20.60] - 2025-10-03
 
 - **Added** new _utility types_ `ArrayToTuple<T[]>` and `Tuple<T>`.
-- Updated
+- **Updated** the _return type_ of `extractObjectKeys`, now it returns properly typed tuple.
+- **Updated** _query string parser_ utilities to receive _generic return type_.
+- **Added** new utility to parse _literal query string_ `literalQueryStringToObject`.
 
 ## [4.20.56] - 2025-10-02